Created CreateListDialog component
